When string cannot be converted into a symbol, log an error message
Often users will try to use a string in place of a symbol when the Symbol has not been registered using a method like AddSecurity, AddCrypto, AddEquity, etc. When this happens Lean fails to implicity convert the string into a symbol object. This PR adds an error message when Lean is unable to convert the ticker into a Symbol obkect. This should add more context when users try to say, add an EMA using a string which has not been registered in the Algorithm as a symbol. This will result in a empty symbol being used, which will get rejected by Lean with a cryptic error message. The error message is cryptic precisely because Lean cannot determine what symbol the user meant to use.